GPkg.jl removed GDAL for now as it messes up on my setup, havent got it working with my julia environment.
This means KML stuff is removed temporarily for now as well.

pushed my sorta working implementation, it hasn't passed the E2E test of read -> write -> read -> write

Also doesn't load the geometry into QGIS on database connection, this will require some in-depth testing.
