+++ Project.toml
@@ -1,10 +1,9 @@
 name = "GeoIO"
 uuid = "f5a160d5-e41d-4189-8b61-d57781c419e3"
 authors = ["Elias Carvalho <eliascarvdev@gmail.com> and contributors"]
-version = "1.20.7"
+version = "1.20.5"
 
 [deps]
-ArchGDAL = "c9ce4bd3-c3d5-55b8-8973-c0e20141b8c3"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 CommonDataModel = "1fbeeb36-5f17-413c-809b-666fb144f157"
@@ -27,6 +26,7 @@ PlyIO = "42171d58-473b-503a-8d5f-782019eb09ec"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 ReadVTK = "dc215faf-f008-4882-a9f7-a79a826fadc3"
+SQLite = "0aa819cd-b072-5ff4-a722-6bc24af294d9"
 Shapefile = "8e980c4a-a4fe-5da2-b3a7-4b4b0353a2f4"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
@@ -36,7 +36,6 @@ VTKBase = "4004b06d-e244-455f-a6ce-a5f9919cc534"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 
 [compat]
-ArchGDAL = "0.10"
 CSV = "0.10"
 Colors = "0.12, 0.13"
 CommonDataModel = "0.2, 0.3"
@@ -48,7 +47,7 @@ GeoFormatTypes = "0.4"
 GeoInterface = "1.0"
 GeoJSON = "0.8"
 GeoParquet = "0.1, 0.2, 0.3"
-GeoTIFF = "0.5"
+GeoTIFF = "0.5.0"
 GeoTables = "1.24.0"
 GslibIO = "1.5"
 ImageIO = "0.6"
