+++ src/GeoIO.jl
@@ -47,11 +47,10 @@ import GeoTIFF
 # GIS formats
 import Shapefile as SHP
 import GeoJSON as GJS
-import ArchGDAL as AG
 import GeoParquet as GPQ
 import GeoInterface as GI
 import GeoFormatTypes as GFT
-import ArchGDAL.GDAL
+
 
 # PROJJSON CRS
 import JSON3
@@ -72,12 +71,12 @@ const CDMEXTS = [".grib", ".nc"]
 const FORMATS = [
   (extension=".csv", load="CSV.jl", save="CSV.jl"),
   (extension=".geojson", load="GeoJSON.jl", save="GeoJSON.jl"),
-  (extension=".gpkg", load="ArchGDAL.jl", save="ArchGDAL.jl"),
+#  (extension=".gpkg", load="GeoIO.jl", save="GeoIO.jl"),
   (extension=".grib", load="GRIBDatasets.jl", save=""),
   (extension=".gslib", load="GslibIO.jl", save="GslibIO.jl"),
   (extension=".jpeg", load="ImageIO.jl", save="ImageIO.jl"),
   (extension=".jpg", load="ImageIO.jl", save="ImageIO.jl"),
-  (extension=".kml", load="ArchGDAL.jl", save=""),
+ # (extension=".kml", load="GeoIO.jl", save="GeoIO.jl"),
   (extension=".msh", load="GeoIO.jl", save="GeoIO.jl"),
   (extension=".nc", load="NCDatasets.jl", save="NCDatasets.jl"),
   (extension=".obj", load="GeoIO.jl", save="GeoIO.jl"),
@@ -124,7 +123,7 @@ include("conversion.jl")
 # extra code for backends
 include("extra/cdm.jl")
 include("extra/csv.jl")
-include("extra/gdal.jl")
+#include("extra/gdal.jl")
 include("extra/geotiff.jl")
 include("extra/gis.jl")
 include("extra/img.jl")
