+++ src/extra/gis.jl
@@ -50,8 +50,6 @@ function giswrite(fname, geotable; warn, kwargs...)
   elseif endswith(fname, ".parquet")
     CRS = crs(domain(geotable))
     GPQ.write(fname, geotable, (:geometry,), projjson(CRS); kwargs...)
-  else # fallback to GDAL
-    agwrite(fname, geotable; kwargs...)
   end
 end
 
@@ -63,9 +61,6 @@ function gistable(fname; layer, numtype, kwargs...)
     return GJS.read(fname; numbertype=numtype, kwargs...)
   elseif endswith(fname, ".parquet")
     return GPQ.read(fname; kwargs...)
-  else # fallback to GDAL
-    data = AG.read(fname; kwargs...)
-    return AG.getlayer(data, layer - 1)
   end
 end
