+++ src/precompile.jl
@@ -12,7 +12,7 @@ using PrecompileTools
     load(joinpath(datadir, "points.parquet"))
     load(joinpath(datadir, "lines.parquet"))
     load(joinpath(datadir, "polygons.parquet"))
-    load(joinpath(datadir, "field.kml"))
+#    load(joinpath(datadir, "field.kml"))
     load(joinpath(datadir, "points.gpkg"))
     load(joinpath(datadir, "lines.gpkg"))
     load(joinpath(datadir, "polygons.gpkg"))
