+++ test/runtests.jl
@@ -14,7 +14,6 @@ using JSONSchema
 import ReadVTK
 import GeoInterface as GI
 import Shapefile as SHP
-import ArchGDAL as AG
 import GeoJSON as GJS
 
 using Unitful: cm
